move project into freqtrade/
