+++ freqtrade/analyze.py
