+++ freqtrade/misc.py
