+++ freqtrade/persistence.py
@@ -5,11 +5,9 @@ from sqlalchemy import Boolean, Column, DateTime, Float, Integer, String, create
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm.scoping import scoped_session
 from sqlalchemy.orm.session import sessionmaker
-
 from sqlalchemy.types import Enum
 
-import exchange
-
+from freqtrade import exchange
 
 _CONF = {}
