+++ freqtrade/tests/test_backtesting.py
@@ -1,14 +1,17 @@
 # pragma pylint: disable=missing-docstring
-import unittest
-from unittest.mock import patch
-import os
 import json
 import logging
+import os
+import unittest
+from unittest.mock import patch
+
 import arrow
 from pandas import DataFrame
-from analyze import analyze_ticker
-from persistence import Trade
-from main import should_sell
+
+from freqtrade.analyze import analyze_ticker
+from freqtrade.main import should_sell
+from freqtrade.persistence import Trade
+
 
 def print_results(results):
     print('Made {} buys. Average profit {:.2f}%. Total profit was {:.3f}. Average duration {:.1f} mins.'.format(
@@ -37,12 +40,12 @@ class TestMain(unittest.TestCase):
     @unittest.skipIf(not os.environ.get('BACKTEST', False), "slow, should be run manually")
     def test_backtest(self):
         trades = []
-        with patch.dict('main._CONF', self.conf):
+        with patch.dict('freqtrade.main._CONF', self.conf):
             for pair in self.pairs:
-                with open('test/testdata/'+pair+'.json') as data_file:
+                with open('testdata/'+pair+'.json') as data_file:
                     data = json.load(data_file)
 
-                    with patch('analyze.get_ticker', return_value=data):
+                    with patch('freqtrade.analyze.get_ticker', return_value=data):
                         with patch('arrow.utcnow', return_value=arrow.get('2017-08-20T14:50:00')):
                             ticker = analyze_ticker(pair)
                             # for each buy point
