+++ freqtrade/tests/testdata/btc-edg.json
