+++ freqtrade/tests/testdata/btc-etc.json
