+++ freqtrade/tests/testdata/btc-eth.json
