+++ freqtrade/tests/testdata/btc-ltc.json
