+++ freqtrade/tests/testdata/btc-mtl.json
