+++ freqtrade/tests/testdata/btc-neo.json
