+++ freqtrade/tests/testdata/btc-omg.json
