+++ freqtrade/tests/testdata/btc-pay.json
