+++ freqtrade/tests/testdata/btc-pivx.json
