+++ freqtrade/tests/testdata/btc-qtum.json
